Modify autoconfig for eBay install
Autoconfig system will need tobe reworked if user tries to install market from the OC install page - this is because the container application (OpenBay Pro) has not been installed with the DB tables or Events etc. 

OBP module and eBay marketplace need to be installed to use the autoconfig/install feature at the end of the registration flow.